refactor: 全局迁移至 int64 ID 以兼容 Snowflake 算法

+++ bookstore-go/repository/favorite.go
@@ -15,7 +15,7 @@ func NewFavoriteDAO() *FavoriteDAO {
 	return &FavoriteDAO{db: global.GetDB()}
 }
 
-func (f *FavoriteDAO) AddFavorite(userID, bookID int) error {
+func (f *FavoriteDAO) AddFavorite(userID, bookID int64) error {
 	favorite := &model.Favorite{
 		UserID: userID,
 		BookID: bookID,
@@ -27,7 +27,7 @@ func (f *FavoriteDAO) AddFavorite(userID, bookID int) error {
 	return nil
 }
 
-func (f *FavoriteDAO) RemoveFavorite(userID, bookID int) error {
+func (f *FavoriteDAO) RemoveFavorite(userID, bookID int64) error {
 
 	err := f.db.Debug().Where("user_id = ? AND book_id = ?", userID, bookID).Delete(&model.Favorite{}).Error
 	if err != nil {
@@ -36,7 +36,7 @@ func (f *FavoriteDAO) RemoveFavorite(userID, bookID int) error {
 	return nil
 }
 
-func (f *FavoriteDAO) GetUserFavorite(userID int) ([]*model.Favorite, error) {
+func (f *FavoriteDAO) GetUserFavorite(userID int64) ([]*model.Favorite, error) {
 	var fav []*model.Favorite
 	err := f.db.Debug().Preload("Book").Where("user_id = ?", userID).Find(&fav).Error
 	if err != nil {
@@ -45,7 +45,7 @@ func (f *FavoriteDAO) GetUserFavorite(userID int) ([]*model.Favorite, error) {
 	return fav, err
 }
 
-func (f *FavoriteDAO) GetUserFavoriteCount(userID int) (int64, error) {
+func (f *FavoriteDAO) GetUserFavoriteCount(userID int64) (int64, error) {
 	var count int64
 	err := f.db.Model(&model.Favorite{}).Debug().Where("user_id = ?", userID).Count(&count).Error
 	if err != nil {
@@ -54,7 +54,7 @@ func (f *FavoriteDAO) GetUserFavoriteCount(userID int) (int64, error) {
 	return count, nil
 }
 
-func (f *FavoriteDAO) CheckFavorite(userID, bookID int) (bool, error) {
+func (f *FavoriteDAO) CheckFavorite(userID, bookID int64) (bool, error) {
 	var count int64
 	err := f.db.Model(&model.Favorite{}).Debug().Where("user_id = ? AND book_id = ?", userID, bookID).Count(&count).Error
 	if err != nil {
